Update host_path assignment in volumes block
Changed the method of assigning the 'host_path' variable within the 'volumes' block in main.tf. Now, 'data_path' is incorporated using the 'pathexpand' function to allow users to define a custom data storage path. In the absence of a user-defined 'data_path', the system will default to the original path.